requirements.txt
flask
shortuuid
python-dotenv
requests
pyperclip
plyer
flask_image_resizer
